requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.0
psycopg2-binary>=2.9
drf-spectacular>=0.27
Pillow>=9.0
celery>=5.2
redis>=4.0
pytest>=7.0
pytest-django>=4.5
python-dotenv>=1.0
gunicorn>=20.1
whitenoise>=6.0
dj-database-url>=1.0
django-storages[boto3]>=1.13
boto3>=1.26
